refactored book_settings into it's own file; update book_config and book.py to use new book_settings file; added {christian_name} {surname} as an alias as well
